+++ gulpfile.js
@@ -12,7 +12,6 @@ const jsFiles = [
   './node_modules/jquery/dist/jquery.js',
   './node_modules/bootstrap/dist/js/bootstrap.js',
   './node_modules/chart.js/dist/Chart.js',
-  './node_modules/moment/moment.js',
   './node_modules/numeral/numeral.js',
   './node_modules/underscore/underscore.js',
   './src/js/helpers.js',